+++ OpenGLProject/OpenGLProject/DirectionalLight.cpp
@@ -21,7 +21,7 @@ DirectionalLight::DirectionalLight(const glm::vec3& inDirection, const Reflectio
 
 void DirectionalLight::Store(const std::vector<uint32_t>& entitiesShadersIDs)
 {
-	ubo = new UniformBuffer(entitiesShadersIDs, "directionalLightParams", static_cast<size_t>(4 * Utils::vec4Size + Utils::scalarSize));
+	ubo = new UniformBuffer(entitiesShadersIDs, "directionalLightParams", 4 * Utils::vec4Size + Utils::scalarSize);
 	ubo->InitSubData({
 		{ static_cast<const void*>(glm::value_ptr(direction)), Utils::vec4Size },
 		{ static_cast<const void*>(glm::value_ptr(reflectionParams.ambient)), Utils::vec4Size },
